+++ samples/ping/src/Ping.kt
@@ -3,6 +3,7 @@ package samples
 import com.serebit.strife.bot
 import com.serebit.strife.entities.UnicodeEmoji
 import com.serebit.strife.entities.reply
+import com.serebit.strife.onGuildRoleUpdate
 import com.serebit.strife.onMessageCreate
 import com.serebit.strife.onReady