+++ CHANGELOG.md
@@ -11,7 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - `GuildRole` now has a `guildId` property, alongside `getGuild` and `delete` functions
 - Add remaining shorthand event DSL
 - Add enum `Type` to `Channel` interface
-- Add `AuditLogs` with `getHistory` flow function.
+- Add `Guild#getSelfMember` extension function
+- Add `GuildBan` class and`Guild#getBans` function
 - [#4](https://gitlab.com/serebit/strife/issues/4) Implement all remaining event types
 - [#17](https://gitlab.com/serebit/strife/issues/17) Add global markdown extensions on String
 - [#20](https://gitlab.com/serebit/strife/issues/20) Add terminable event listeners, which are removed from the client when their task is successful
@@ -38,6 +39,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
         - setColor
         - setHoisted, hoist, unhoist
         - setMentionable
+        - Set position, raise, lower
     - GuildChannel
         - get & create Invites
     - GuildIntegration (new)